pyproject.toml
[tool.poetry]
name = "taskmanagerapi"
version = "0.1.0"
description = ""
authors = ["md <marceldiomare@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
aiohttp = "^3.11.1"
asyncpg = "^0.30.0"
pydantic = {extras = ["dotenv"], version = "^2.9.2"}
aioredis = "^2.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
uvicorn = "^0.32.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
